fix seed and change type data spk
spk from integer to decimal

+++ database/migrations/2023_10_02_231424_api_datas.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->integer('ipk');
             $table->integer('alpha');
             $table->integer('sertifikat');
-            $table->integer('spk');
+            $table->decimal('spk');
             $table->string('username');
             $table->string('password');
             $table->timestamps();
